legacy/src/train/requirements.txt
pandas>=2.0.0
pytest>=7.3.1
numpy>=1.26.0
scikit-learn>=1.3.0
xgboost>=1.7.5
mlflow>=2.3.1
pyyaml>=6.0
dvc>=2.28.0
